Improve sockets functions and console commands
Add functions for filtering all the clients (socket.id) for a specific user, retrieving all the connected clients properly. Found a bug with:
` socket.on("disconnect", function(){});` not invoking when the client refreshes the page, thus flooding the clients counter in the sockets.js file. This gives back wrong information that no-longer exsisting sockets are still connected. This problem was avoided by adding a "user" object to each socket and filtering the information from `io.sockets.sockets` when needed.

Also added commands that can be invoked by typing in the console. The supported commands at the moment are:
* `show users` displays a list of the connected users along with the socket.id
* `send <message>` sends a message to all the connected sockets
* `send to <email> <message>` sends a message to all the connected clients of a user, specified by email
More commands can be added later but reallocating them to a different file should be considered.