Make sample game work
The sample ping-pong game is now working, but is not yet in demo state.
The game can be played by two players. It is started from the lobby when one user invites another to a game, by selecting the "options" icon next to the user, and clicking on "Invite to game". The other user then has the option to accept the invitation or refuse it.

There is yet to be done to optimise the multiplayer module and the canvas rendering